Move redirects methods to trait
